requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.0
konlpy==0.6.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0

